moved form code into components
